rf: command structure changes
