+++ Jayuwoki/src/main/java/dad/controllers/MainController.java
@@ -43,6 +43,7 @@ public class MainController implements Initializable {
     private final ContactController contactController = new ContactController();
     private final ConnectController connectController = new ConnectController();
     private final AboutController aboutController = new AboutController();
+    private final SettingsController settingsController = new SettingsController();
 
     @FXML
     private VBox aboutBox;
@@ -149,6 +150,8 @@ public class MainController implements Initializable {
                 customTitleBar.onCloseWindow();
             }
         });
+
+        // bind setting checkbox
     }
 
     private void popOverMap() {
@@ -295,7 +298,6 @@ public class MainController implements Initializable {
 
     @FXML
     void onSettingsAction(ActionEvent event) {
-        SettingsController settingsController = new SettingsController();
         contentPane.getChildren().setAll(settingsController.getRoot());
     }
